% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaCalendarDate2rPOSIXlt.R
\name{javaCalendarDate2rPOSIXlt}
\alias{javaCalendarDate2rPOSIXlt}
\title{Java calendar dates to R POSIXlt dates}
\usage{
javaCalendarDate2rPOSIXlt(javaCalendarDate)
}
\arguments{
\item{javaCalendarDate}{A \sQuote{ucar.nc2.time.CalendarDate} object or array 
containing N of them.}
}
\value{
A vector of \code{POSIXlt} dates in R
}
\description{
Converts objects of the Java class \sQuote{CalendarDate} to \link[base]{POSIXlt} in R
}
\references{
\url{http://www.unidata.ucar.edu/software/thredds/v4.3/netcdf-java/v4.3/javadoc/index.html}
}
\author{
J. Bedia
}
\keyword{internal}
