% Generated by roxygen2 (4.0.1): do not edit by hand
\name{javaString2rChar}
\alias{javaString2rChar}
\title{Java string dates to R character}
\usage{
javaString2rChar(javaString)
}
\arguments{
\item{javaString}{A \sQuote{java/lang/String} object or array}
}
\value{
A character vector in R
}
\description{
Converts objects of the Java class \sQuote{java/lang/String} to character in R
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

