% Generated by roxygen2 (4.0.1): do not edit by hand
\name{javaCalendarDate2rPOSIXlt}
\alias{javaCalendarDate2rPOSIXlt}
\title{Java calendar dates to R POSIXlt dates}
\usage{
javaCalendarDate2rPOSIXlt(javaCalendarDate)
}
\arguments{
\item{javaCalendarDate}{A \sQuote{ucar.nc2.time.CalendarDate} object or array
containing N of them.}
}
\value{
A vector of \code{POSIXlt} dates in R
}
\description{
Converts objects of the Java class \sQuote{CalendarDate} to \link[base]{POSIXlt} in R
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\references{
\url{https://www.unidata.ucar.edu/software/thredds/current/netcdf-java/v4.3/javadoc/ucar/nc2/time/CalendarDate.html}
}
\keyword{internal}

